backend/requirements.txt
# ═══════════════════════════════════════════════════════
# NIX AI Backend — Python Dependencies
# ═══════════════════════════════════════════════════════

# ── Core Framework ──
fastapi==0.115.0
uvicorn[standard]==0.30.0
mangum==0.18.0

# ── Data Validation ──
pydantic==2.11.3
python-multipart==0.0.9

# ── AWS SDK ──
boto3>=1.34.72,<2.0.0

# ── Auth ──
python-jose[cryptography]==3.3.0

# ── PDF Processing ──
pypdf>=4.0.0

# ── HTTP Client (for tests / integrations) ──
requests==2.32.0
httpx==0.27.0

# ═══════════════════════════════════════════════════════
# Dev / Test Dependencies (not needed in Lambda)
# ═══════════════════════════════════════════════════════
python-dotenv==1.0.1

# pytest==8.2.0
# pytest-asyncio==0.23.0
# moto[dynamodb,s3,sqs]==5.0.0